requirements.txt
filetype
python-magic
chardet
pillow
opencv-python
PyPDF2
pdfplumber
python-docx
openpyxl
pandas
ffmpeg-python
moviepy
pydub